fix: forwarder name change to include connector
The forwarder name needs to include the source connector, since multiple
source connectors can foward to the same network.

Also introduce a change in the naming of the connector/networks. Since
Agilicus network names need to be RFC1123, utilize the ansible_host to override
the host with an ip address. This is only relevant when using ip
addresses directly instead of hostnames.